backend/requirements.txt
Pillow==10.4.0
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9